+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.Tasklist;
+import duke.Ui;
+import duke.exceptions.NoSuchTaskException;
+
 import java.io.IOException;
 
 public abstract class Command {
