+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,10 +1,18 @@
+package duke.commands;
+
+import duke.Tasklist;
+import duke.exceptions.NoSuchTaskException;
+import duke.tasks.Task;
+import duke.Ui;
+import duke.Storage;
+
 import java.io.IOException;
 
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
 
     private int index;
 
-    public DoneCommand(int index) {
+    public DeleteCommand(int index) {
         this.index = index;
     }
 
@@ -12,11 +20,12 @@ public class DoneCommand extends Command {
     public void execute(Tasklist tasklist, Ui ui, Storage storage) throws NoSuchTaskException {
         try {
             Task temp = tasklist.getTask(index);
-            tasklist.doneTask(index);
-            ui.showDoneMessage(temp, tasklist);
+            tasklist.deleteTask(index);
+            ui.showDeleteMessage(temp, tasklist);
             storage.writeToFile(tasklist);
         } catch (NoSuchTaskException | IOException e) {
             throw new NoSuchTaskException("Task index is out of bounds");
         }
+
     }
 }
