+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,10 +1,18 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.Tasklist;
+import duke.Ui;
+import duke.exceptions.NoSuchTaskException;
+import duke.tasks.Task;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class DoneCommand extends Command {
 
     private int index;
 
-    public DeleteCommand(int index) {
+    public DoneCommand(int index) {
         this.index = index;
     }
 
@@ -12,12 +20,11 @@ public class DeleteCommand extends Command {
     public void execute(Tasklist tasklist, Ui ui, Storage storage) throws NoSuchTaskException {
         try {
             Task temp = tasklist.getTask(index);
-            tasklist.deleteTask(index);
-            ui.showDeleteMessage(temp, tasklist);
+            tasklist.doneTask(index);
+            ui.showDoneMessage(temp, tasklist);
             storage.writeToFile(tasklist);
         } catch (NoSuchTaskException | IOException e) {
             throw new NoSuchTaskException("Task index is out of bounds");
         }
-
     }
 }
