+++ src/main/java/duke/exceptions/NoSuchCommandException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class NoSuchCommandException extends Exception{
     public NoSuchCommandException(String message){
         super(message);
