+++ src/main/java/duke/exceptions/NoSuchTaskException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class NoSuchTaskException extends Exception{
     public NoSuchTaskException (String message){
         super(message);
